authorization implemented

+++ src/services/notes/controller/note-controller.js
@@ -1,14 +1,16 @@
-import { nanoid } from 'nanoid';
-import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
+import { AuthorizationError, InvariantError, NotFoundError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
 import NoteRepositories from '../repositories/note-repositories.js';
 
 export const createNote = async (req, res, next) => {
   const { title = 'untitled', tags, body } = req.validated;
+  const { id: owner } = req.user;
+
   const note = await NoteRepositories.createNote({
     title,
     body,
-    tags
+    tags,
+    owner,
   });
 
   if (!note) {
@@ -19,24 +21,41 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
-  return response(res, 200, 'success', { notes: notes });
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
+  return response(res, 200, 'Catatan sukses ditampilkan', { notes: notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.getNoteById(id);
 
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
 
-  return response(res, 200, 'Catatan sukses ditampilkan', note);
+  return response(res, 200, 'Catatan sukses ditampilkan', { note });
 };
 
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, tags, body } = req.validated;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.editNote({
     id,
     title,
@@ -53,12 +72,19 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const deletedNote = await NoteRepositories.deleteNote(id);
 
   if (!deletedNote) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
 
-  notes.splice(index, 1);
   return response(res, 200, 'Catatan berhasil dihapus', deletedNote);
 };
